requirements.txt
flask==3.1.2
opencv-python-headless==4.11.0.86
numpy==2.3.3
pillow==11.3.0
gunicorn==23.0.0